+++ src/tr1/game/inventory_ring/control.c
@@ -816,9 +816,6 @@ INV_RING *InvRing_Open(const INVENTORY_MODE mode)
     if (!g_Config.audio.enable_music_in_inventory && mode != INV_TITLE_MODE) {
         Music_Pause();
         Sound_PauseAll();
-    } else {
-        Sound_ResetAmbient();
-        Sound_UpdateEffects();
     }
 
     INV_RING *const ring = Memory_Alloc(sizeof(INV_RING));